user messages and message editor

+++ src/main/java/com/example/repo/MessageRepo.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MessageRepo extends CrudRepository<Message, Integer> {
+public interface MessageRepo extends CrudRepository<Message, Long> {
     List<Message> findAllByTag(String tag);
 }
